Fix length column type mismatch after domain-objects merge
The Dicks.length column became INT8 (bigint) in #116, but the
fetch_length_locked helper added by the over-loaning fix still
decoded it as i32, panicking at runtime (INT4 vs INT8). Return
i64 and update the offline .sqlx cache accordingly.

+++ src/repo/loans.rs
@@ -94,7 +94,7 @@ impl Loans {
     }
 }
 
-async fn fetch_length_locked(tx: &mut Transaction<'_, Postgres>, uid: UserId, chat_internal_id: InternalChatId) -> anyhow::Result<Option<i32>> {
+async fn fetch_length_locked(tx: &mut Transaction<'_, Postgres>, uid: UserId, chat_internal_id: InternalChatId) -> anyhow::Result<Option<i64>> {
     sqlx::query_scalar!("SELECT length FROM Dicks WHERE chat_id = $1 AND uid = $2 FOR UPDATE",
             chat_internal_id as InternalChatId, uid as UserId)
         .fetch_optional(&mut **tx)
